+++ ChangeLog.md
@@ -1,6 +1,18 @@
 ------------------
 CHANGE LOG
 -------------------
+
+--------
+201702XX
+--------
+
+1.6.8
+
+Fixed tests to reference TLS certs in test directory not in /etc/pki/..
+Added posix timestamp that works for both python 2.7+ and is aware
+Added optilnal aware keyword parameter to iso8601
+fixed tuuid so includes microseconds
+
 --------
 20170110
 --------